+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-virtual-list
 
+## 0.0.5
+
+### Patch Changes
+
+-   modified scrollTo, revealed scrollPos, added docs & prop-types, upgraded deps
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes